third_party/blink/web_tests/wpt_internal/shared_storage/run-operation.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<body>
<script>
'use strict';

promise_test(() => {
  return sharedStorage.run("operation1", {keepAlive: true});
}, 'run()');

promise_test(() => {
  return sharedStorage.run("operation1",
                           {data: {'custom-key': 'custom-value'},
                            keepAlive: true});
}, 'run() with custom data');

promise_test(async t => {
  try {
    const blob = new Blob(["abc"], {type: 'text/plain'});
    await sharedStorage.run("operation1", {data: blob});
  } catch (e) {
    assert_equals(e.name, 'Error');
    assert_equals(e.message, '#<Blob> could not be cloned.');
    return;
  }
  assert_unreached("did not reject");
}, 'run() with unsupported serializable type');

</script>
</body>
